modularize components and helpers
