+++ src/components/HelperControls/index.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const HelperControls = styled.div`
+  display: flex;
+  margin-top: 1em;
+`
+
+export default HelperControls;
