+++ src/components/Mask/index.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const Mask = styled.div`
+  background-color: rgba(0,0,0,.7);
+  width: 100%;
+  left: 0;
+  top: 0;
+  height: 100%;
+  position: fixed;
+  z-index: 99999;
+`
+
+export default Mask;
