+++ src/components/Navigation/index.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const Navigation = styled.nav`
+  counter-reset: dot;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex: 1;
+`
+
+export default Navigation;
