+++ src/utils/getNodeRect.js
@@ -1,0 +1,10 @@
+import pick from 'lodash.pick';
+
+const getNodeRect = node => {
+  return pick(node.getBoundingClientRect(), [
+    'top', 'right', 'bottom', 'left',
+    'width', 'height',
+  ])
+};
+
+export default getNodeRect;
