+++ src/utils/isBody.js
@@ -1,0 +1,3 @@
+const isBody = node => node === document.querySelector('body');
+
+export default isBody;
