+++ src/utils/isHoriz.js
@@ -1,0 +1,3 @@
+const isHoriz = pos => /(left|right)/.test(pos);
+
+export default isHoriz;
