+++ src/utils/isOutsideX.js
@@ -1,0 +1,3 @@
+const isOutsideX = (val, windowWidth) => val > windowWidth;
+
+export default isOutsideX;
