+++ src/utils/isOutsideY.js
@@ -1,0 +1,3 @@
+const isOutsideY = (val, windowHeight) => val > windowHeight;
+
+export default isOutsideY;
