+++ src/utils/safe.js
@@ -1,0 +1,3 @@
+const safe = sum => sum < 0 ? 0 : sum;
+
+export default safe;
